[#177458092] refactoring
